Racmacs: R Antigenic Cartography Macros

The acmap data object

The fundamental unit of the Racmacs package is the acmap object, short for Antigenic Cartography MAP. This object contains all the information about an antigenic map. You can read in a new acmap object from a file with the function read.acmap and create a new acmap object within an R session using the acmap function.

Key information associated with each acmap object is summarized in the sections below.

The titer table

Each acmap object is built upon a table of data of measurements the reactivity of a set of different sera against a set of different antigens. Typically these measurements are HI assay measurements, but it is also possible to use other similar assays as input data.

For the table of data it is always assumed that sera form the columns and antigens form the rows, as below. You can get and set titer data with titerTable.

Optimizations

Another key component of the acmap object is a list of optimizations. While acmap objects only have one table of data, they can have many optimizations or none at all.

Each optimization has the following main attributes (see the vignette on optimizing antigenic maps for more details on minimum column bases and stress):

Antigen coordinates, the coordinates of each antigen in this optimization.

Sera coordinates, the coordinates of each serum in this optimization.

Minimum column basis, the minimum column basis assumed when calculating this optimization.

Stress, the stress of this optimization.

Dimensions, the number of dimensions of this optimization.

Plotting styles

The final type of information that is contained in the acmap object is information on point styles when plotting. By altering these attributes you can change the appearance of the antigen and serum points in any maps plotted, the main ones include:

Size

Shape

Fill color

Outline color

See also