Generate a new acmap object

This function generates a new acmap object, the base object for storing map data in the Racmacs package.

acmap(...)

Arguments

table_name
Table name

table
Table of titer data

ag_names
Antigen names

ag_date
Antigen dates

ag_reference
Is antigen a reference virus

sr_names
Sera names

ag_shown
Antigen shown

ag_size
Antigen size

ag_cols_fill
Antigen fill color

ag_cols_outline
Antigen outline color

ag_outline_width
Antigen outline width

ag_rotation
Antigen rotation

ag_aspect
Antigen aspect

ag_shape
Antigen shape

ag_drawing_order
Antigen drawing order

sr_shown
Sera shown

sr_size
Sera size

sr_cols_fill
Sera fill color

sr_cols_outline
Sera outline color

sr_outline_width
Sera outline width

sr_rotation
Sera rotation

sr_aspect
Sera aspect

sr_shape
Sera shape

sr_drawing_order
Sera drawing order

ag_coords
Antigen coordinates

sr_coords
Sera coordinates

comment
Map comment

minimum_column_basis
Map minimum column bases

transformation
Map transformation

colbases
Map column bases

optimizations
A list of optimizations with appropriate attributes

Value

Returns the new acmap object

Details

The fundamental unit of the Racmacs package is the acmap object, short for Antigenic Cartography MAP. This object contains all the information about an antigenic map. You can read in a new acmap object from a file with the function read.acmap and create a new acmap object within an R session using the acmap function.

See also

Examples

# Read in an acmap object from a file
map_file_path <- system.file("extdata/h3map2004.ace", package = "Racmacs")
map <- read.acmap(map_file_path)

# Read in a table of titer data and make a new acmap file from it
titer_file_path <- system.file("extdata/h3map2004_hitable.csv", package = "Racmacs")
titer_data <- read.titerTable(titer_file_path)
map <- acmap.cpp(table = titer_data)