Calculate bootstrap blob data for an antigenic map

This function takes a map for which the function bootstrapMap() has already been applied and draws contour blobs for each point illustrating how point position varies in each bootstrap repeat. The blobs are calculated using kernal density estimates according to these point distribution and drawn so as to encompass a given proportion of this variation according to the parameter conf.level. A conf.level set at 0.95 for example will draw blobs that are calculated to encompass 95% of the positional variation seen in the bootstrap repeats. Note however that the accuracy of these estimates will depend on the number of bootstrap repeats performed, for example whether 100 or 1000 repeats were performed in the initial calculations using boostrapMap().

bootstrapBlobs(map, conf.level = 0.68, smoothing = 6)

Arguments

map
The acmap data object

conf.level
The proportion of positional variation captured by each blob

smoothing
The amount of smoothing to perform when performing the kernel density estimate, larger equates to more smoothing

Value

Returns an acmap object that will then show the corresponding bootstrap blobs when viewed or plotted.

See also