A simple mesh generator for non-convex regions in n-D space

An unstructured simplex requires a choice of meshpoints (vertex nodes) and a triangulation. This is a simple and short algorithm that improves the quality of a mesh by relocating the meshpoints according to a relaxation scheme of forces in a truss structure. The topology of the truss is reset using Delaunay triangulation. A (sufficiently smooth) user supplied signed distance function (fd) indicates if a given node is inside or outside the region. Points outside the region are projected back to the boundary.

distmeshnd(fdist, fh, h, box, pfix = array(dim = c(0, ncol(box))),
  ptol = 0.001, ttol = 0.1, deltat = 0.1, geps = 0.1 * h,
  deps = sqrt(.Machine$double.eps) * h, max_delauney = 10, ...)

Arguments

fdist
Vectorized signed distance function, for example mesh.dsphere, accepting an m-by-n matrix, where m is arbitrary, as the first argument.

fh
Vectorized function, for example mesh.hunif, that returns desired edge length as a function of position. Accepts an m-by-n matrix, where n is arbitrary, as its first argument.

h
Initial distance between mesh nodes.

box
2-by-n matrix that specifies the bounding box. (See distmesh2d for an example.)

pfix
nfix-by-2 matrix with fixed node positions.

ptol
Algorithm stops when all node movements are smaller than ptol

ttol
Controls how far the points can move (relatively) before a retriangulation with delaunayn.

deltat
Size of the time step in Eulers method.

geps
Tolerance in the geometry evaluations.

deps
Stepsize delta x in numerical derivative computation for distance function.

max_delauney
Maximum delauney

...
parameters that are passed to fdist and fh