Plot an antigenic map using ggplot

Method for plotting an antigenic map as a ggplot object

# S3 method for acmap
ggplot(
  data = NULL,
  mapping = NULL,
  optimization_number = 1,
  xlim = NULL,
  ylim = NULL,
  plot_ags = TRUE,
  plot_sr = TRUE,
  plot_blobs = TRUE,
  plot_hemisphering = TRUE,
  show_procrustes = TRUE,
  show_error_lines = FALSE,
  plot_stress = FALSE,
  indicate_outliers = "arrowheads",
  grid.col = "grey90",
  grid.lwd = 0.5,
  grid.margin.col = "grey50",
  grid.margin.lwd = grid.lwd,
  fill.alpha = 0.8,
  outline.alpha = 0.8,
  padding = 1,
  arrow_angle = 25,
  arrow_length = 0.2,
  margins = rep(0.5, 4),
  ...,
  environment = NULL
)

Arguments

data

The acmap to plot

mapping

Default list of aesthetic mappings to use for plot, not currently used

optimization_number

The optimization number to plot

xlim

optional x axis limits

ylim

optional y axis limits

plot_ags

logical, should antigens be plotted

plot_sr

logical, should antigens be plotted

plot_blobs

logical, should stress blobs be plotted if present

plot_hemisphering

logical, should hemisphering points be indicated, if tested for already with checkHemisphering() (and if present)

show_procrustes

logical, should procrustes lines be shown, if present

show_error_lines

logical, should error lines be drawn

plot_stress

logical, should map stress be plotted in lower left corner

indicate_outliers

how should points outside the plotting region be indicated, either FALSE, for not shown, or "arrowheads" for small arrowheads like in the viewer.

grid.col

grid line color

grid.lwd

grid line width

grid.margin.col

grid margin color

grid.margin.lwd

grid margin line width

fill.alpha

alpha for point fill

outline.alpha

alpha for point outline

padding

padding at limits of the antigenic map, ignored if xlim or ylim set explicitly

arrow_angle

angle of arrow heads drawn for procrustes lines

arrow_length

length of arrow heads drawn for procrustes lines in cm

margins

margins in inches for the plot

...

additional arguments, not used

environment

not used

Value

Returns the ggplot plot

See also