Plot an antigenic map using the grid system

Plot an antigenic map using the 'grid' plotting system, this is harder to edit afterwards but allows maps with custom point shapes like "EGG" and "UGLYEGG" to be plotted, or maps where point aspect or rotation has been altered.

grid.plot.acmap(
  map,
  optimization_number = 1,
  xlim = NULL,
  ylim = NULL,
  plot_ags = TRUE,
  plot_sr = TRUE,
  plot_labels = FALSE,
  plot_blobs = TRUE,
  plot_hemisphering = TRUE,
  grid.col = "grey90",
  grid.margin.col = grid.col,
  fill.alpha = 0.8,
  outline.alpha = 0.8,
  padding = 1,
  cex = 1
)

Arguments

map
The acmap to plot

optimization_number
The optimization number to plot

xlim
optional x axis limits

ylim
optional y axis limits

plot_ags
logical, should antigens be plotted

plot_sr
logical, should antigens be plotted

plot_labels
logical, should point labels be plotted

plot_blobs
logical, should stress blobs be plotted if present

grid.col
grid line color

grid.margin.col
grid margin color

fill.alpha
alpha for point fill

outline.alpha
alpha for point outline

padding
padding at limits of the antigenic map, ignored if xlim or ylim set explicitly

cex
point size expansion factor

See also