Getting and setting point coordinates

Getting and setting of antigen and serum coordinates in a map optimization run (by default the currently selected one).

agCoords(map, optimization_number = NULL, .name = TRUE)

srCoords(map, optimization_number = NULL, .name = TRUE)

agCoords(map, optimization_number = NULL) <- value

srCoords(map, optimization_number = NULL) <- value

Arguments

map
The acmap object

optimization_number
The optimization number from which to get / set the coordinates

.name
Should the row names of the coordinates be named according to the antigen / serum names

Details

These functions get and set point coordinates in a map. By default these coordinates refer to the currently selected optimization run, unless otherwise specified through the `optimization_number` argument.

99% of the time `agCoords()` and `serumCoords()` are the functions you'll want to use but you should note that the outputs are actually the map base coordinates after the transformation and translation associated with the optimization run has been applied (see `mapTransformation()` and `mapTranslation()` for more details). When you set the antigen or serum coordinates through these functions, the transformed coordinates are "baked" in and the map transformation and translation are reset. Consequently if you want to apply a transformation to all coordinates generally, you are better off modifying the map translation and transformation directly, as is done by functions like `rotateMap()` and `translateMap()`.

See also