Viewing racmap objects

View a racmap object in the interactive viewer.

# S3 method for class 'acmap'
view(
  x,
  optimization_number = 1,
  ...,
  .jsCode = NULL,
  .jsData = NULL,
  select_ags = NULL,
  select_sr = NULL,
  show_procrustes = NULL,
  show_diagnostics = NULL,
  num_optimizations = 1,
  options = list()
)

Arguments

x

The acmap data object

optimization_number

The optimization number to view

...

Additional arguments to be passed to RacViewer()

.jsCode

Additional javascript code to be run after map has been loaded and rendered

.jsData

Any data to supply to the .jsCode function

select_ags

A vector of antigen indices to select in the plot

select_sr

A vector of serum indices to select in the plot

show_procrustes

If the map contains procrustes information, should procrustes lines be shown by default?

show_diagnostics

If the map contains diagnostics information like stress blobs or hemisphering, should it be shown by default?

num_optimizations

Number of optimization runs to send to the viewer for inclusion in the "optimizations" pane.

options

A named list of viewer options to pass to RacViewer.options()

Value

Returns an htmlwidget object

See also