Set acmap optimization options

This function facilitates setting options for the acmap optimizer process by returning a list of option settings.

RacOptimizer.options(
  dim_annealing = FALSE,
  method = "L-BFGS",
  maxit = 1000,
  num_basis = 10,
  armijo_constant = 1e-04,
  wolfe = 0.9,
  min_gradient_norm = 1e-06,
  factr = 1e-15,
  max_line_search_trials = 50,
  min_step = 1e-20,
  max_step = 1e+20,
  num_cores = getOption("RacOptimizer.num_cores"),
  report_progress = NULL,
  ignore_disconnected = FALSE,
  progress_bar_length = options()$width
)

Arguments

dim_annealing

Should dimensional annealing be performed

method

The optimization method to use

maxit

The maximum number of iterations to use in the optimizer

num_basis

Number of memory points to be stored (default 10).

armijo_constant

Controls the accuracy of the line search routine for determining the Armijo condition.

wolfe

Parameter for detecting the Wolfe condition.

min_gradient_norm

Minimum gradient norm required to continue the optimization.

factr

Minimum relative function value decrease to continue the optimization.

max_line_search_trials

The maximum number of trials for the line search (before giving up).

min_step

The minimum step of the line search.

max_step

The maximum step of the line search.

num_cores

The number of cores to run in parallel when running optimizations

report_progress

Should progress be reported

ignore_disconnected

Should the check for disconnected points be skipped

progress_bar_length

Progress bar length when progress is reported

Value

Returns a named list of optimizer options

Details

For more details, for example on "dimensional annealing" see vignette("intro-to-antigenic-cartography"). For details on optimizer settings like maxit see the underlying optimizer documentation at ensmallen.org.

See also