Set acmap optimization options

This function facilitates setting options for the acmap optimizer process by returning a list of option settings.

RacOptimizer.options(
  dim_annealing = FALSE,
  method = "L-BFGS",
  maxit = 1000,
  num_cores = parallel::detectCores(),
  report_progress = NULL,
  progress_bar_length = options()$width
)

Arguments

dim_annealing
Should dimensional annealing be performed

method
The optimization method to use

maxit
The maximum number of iterations to use in the optimizer

num_cores
The number of cores to run in parallel

report_progress
Should progress be reported

progress_bar_length
Progress bar length when progress is reported

Value

Returns a named list of optimizer options

Details

For more details, for example on "dimensional annealing" see vignette("intro-to-antigenic-cartography"). For details on optimizer settings like maxit see the underlying optimizer documentation at ensmallen.org.

See also