Generate a new acmap object

This function generates a new acmap object, the base object for storing map data in the Racmacs package.

acmap(
  ag_names = NULL,
  sr_names = NULL,
  titer_table = NULL,
  ag_coords = NULL,
  sr_coords = NULL,
  check_duplicates = TRUE,
  ...
)

Arguments

ag_names

Antigen names

sr_names

Sera names

titer_table

Table of titer data

ag_coords

Antigenic coordinates for an optimization run record (optional)

sr_coords

Sera coordinates for an optimization run record (optional)

check_duplicates

Issue a warning if duplicate antigen or sera names are found

...

Further arguments passed to addOptimization()

Value

Returns the new acmap object

Details

The fundamental unit of the Racmacs package is the acmap object, short for Antigenic Cartography MAP. This object contains all the information about an antigenic map. You can read in a new acmap object from a file with the function read.acmap() and create a new acmap object within an R session using the acmap() function.

See also